RAB-584: Rename model name in order to having the following error TypeError: Routing.setRoutingData is not a function
