+++ front-packages/shared/src/microfrontend/model/queryParam.unit.ts
@@ -1,4 +1,4 @@
-import {createQueryParam} from './routing';
+import {createQueryParam} from './queryParam';
 
 test('it returns empty query parameter when no parameter is given', () => {
   expect(createQueryParam()).toEqual('');
